Extract subgraph resolving logic to separate class
